refactor simulation_settings to sim_settings
