+++ bim2sim/kernel/ifc_file.py
@@ -1,4 +1,3 @@
-from enum import Enum
 import logging
 
 from pathlib import Path
