+++ bim2sim/plugins/PluginAixLib/bim2sim_aixlib/__init__.py
@@ -5,7 +5,7 @@ from bim2sim.export.modelica import standardlibrary
 from bim2sim.plugins import Plugin
 from bim2sim.plugins.PluginAixLib.bim2sim_aixlib.models import AixLib
 from bim2sim.task import base, common, hvac
-from bim2sim.simulation_settings import PlantSimSettings
+from bim2sim.sim_settings import PlantSimSettings
 
 
 class LoadLibrariesAixLib(base.ITask):
