+++ bim2sim/plugins/PluginCFD/bim2sim_cfd/__init__.py
@@ -4,7 +4,7 @@ Prepares ifc files for CFD simulation
 """
 from bim2sim.plugins import Plugin
 from bim2sim.plugins.PluginCFD.bim2sim_cfd.task.ifc2cfd import RunIFC2CFD
-from bim2sim.simulation_settings import CFDSimSettings
+from bim2sim.sim_settings import CFDSimSettings
 
 
 class PluginCFD(Plugin):
