+++ bim2sim/plugins/PluginCFD/test/integration_test/test_integration.py
@@ -2,7 +2,7 @@ import os
 import unittest
 import warnings
 
-from bim2sim import simulation_settings
+from bim2sim import sim_settings
 from bim2sim.decision.decisionhandler import DebugDecisionHandler
 from bim2sim.utilities.test import IntegrationBase
 from bim2sim.utilities.types import IFCDomain
