+++ bim2sim/plugins/PluginHKESim/bim2sim_hkesim/__init__.py
@@ -5,7 +5,7 @@ Holds logic to run a simulation based on prepared ifc data
 from bim2sim.export.modelica import standardlibrary
 from bim2sim.plugins import Plugin
 from bim2sim.task import base, common, hvac
-from bim2sim.simulation_settings import PlantSimSettings
+from bim2sim.sim_settings import PlantSimSettings
 from .models import HKESim
