+++ bim2sim/plugins/PluginTEASER/test/ap7_testing/ap7_cases.py
@@ -1,7 +1,7 @@
 import unittest
 from pathlib import Path
 
-from bim2sim import simulation_settings
+from bim2sim import sim_settings
 from bim2sim.decision.decisionhandler import DebugDecisionHandler
 from integration_test_setup import \
     IntegrationBaseTeaserInteractive
