+++ bim2sim/sim_settings.py
