+++ test/unit/test_project.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from bim2sim.plugins import Plugin
 from bim2sim.project import Project
-from bim2sim.simulation_settings import PlantSimSettings
+from bim2sim.sim_settings import PlantSimSettings
 from bim2sim.utilities.types import IFCDomain
 
 sample_root = Path(__file__).parent.parent.parent / 'test/TestModels/HVAC'
